(Fix): Update available metrics and rename folktables data server [pv622]
