+++ docker-compose.dev.yml
@@ -171,7 +171,7 @@ services:
     volumes:
       - .:/app
       - venv:/app/.venv
-    command: poetry run python3 -u -m mocks.dataset.mock_server --host 0.0.0.0 --reload
+    command: poetry run python3 -u -m mocks.dataset.folktables_mock_server --host 0.0.0.0 --reload
     working_dir: /app/packages/mocks
     networks:
       - aignostic_network
