+++ packages/api/project.json
@@ -91,7 +91,7 @@
     "mock-dataset": {
       "executor": "@nxlv/python:run-commands",
       "options": {
-        "command": "python3 -m mocks.dataset.mock_server",
+        "command": "python3 -m mocks.dataset.folktables_mock_server",
         "cwd": "packages/mocks"
       }
     },
