+++ packages/metrics/metrics/metrics.py
@@ -61,8 +61,8 @@ task_type_to_metric = {
         "positive_predictive_value",
         "true_positive_rate_difference",
         "explanation_stability_score",
-        # "explanation_sparsity_score",
-        # "explanation_fidelity_score",
+        "explanation_sparsity_score",
+        "explanation_fidelity_score",
         "ood_auroc",
     ],
     "multi_class_classification": [
@@ -74,9 +74,9 @@ task_type_to_metric = {
         # "class_f1_score",
         "f1_score",
         "roc_auc",
-        # "explanation_stability_score",
-        # "explanation_sparsity_score",
-        # "explanation_fidelity_score",
+        "explanation_stability_score",
+        "explanation_sparsity_score",
+        "explanation_fidelity_score",
         "ood_auroc",
     ],
     "regression": [
@@ -84,8 +84,8 @@ task_type_to_metric = {
         "mean_squared_error",
         "r_squared",
         "explanation_stability_score",
-        # "explanation_sparsity_score",
-        # "explanation_fidelity_score",
+        "explanation_sparsity_score",
+        "explanation_fidelity_score",
     ],
     # TODO: Review which metrics really apply
     "text_classification": [
