+++ packages/mocks/tests/app/test_api_keys_unit.py
@@ -1,5 +1,5 @@
 from fastapi.testclient import TestClient
-from mocks.dataset.mock_server import app as data_app
+from mocks.dataset.folktables_mock_server import app as data_app
 from mocks.model.scikit_mock_classifier import app as sk_model_app
 from mocks.model.huggingface_binclassifier import app as hf_model_app
 from mocks.api_utils import MOCK_DATASET_API_KEY, MOCK_MODEL_API_KEY
