+++ swarm/mocks.docker-compose.yml
@@ -7,7 +7,7 @@ services:
 
   folktables_dataset_api:
     image: ghcr.io/aignostic/mocks
-    command: /app/.venv/bin/python -m uvicorn mocks.dataset.mock_server:app --host 0.0.0.0 --port 5010
+    command: /app/.venv/bin/python -m uvicorn mocks.dataset.folktables_mock_server:app --host 0.0.0.0 --port 5010
     working_dir: /app/packages/mocks
     networks:
       - aignostic_network
